overlap/requirements.txt
uuid
qdrant-client
python-dotenv
openai
requests
youtubesearchpython
youtube-transcript-api
beautifulsoup4
streamlit